# src/tools/panel/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IssueCount": {
        "description": "One issue code and how often it occurred.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "count"
        ],
        "title": "IssueCount",
        "type": "object"
      },
...
      "ServiceDetailResponse": {
        "description": "The service page: the registry row plus what only the detail view shows.",
        "properties": {
          "active_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SnapshotResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "docs_changed": {
            "title": "Docs Changed",
            "type": "boolean"
          },
          "docs_ok": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docs Ok"
          },
          "documents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documents"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error At"
          },
          "head_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Commit"
          },
          "initiated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initiated By"
          },
          "interruption": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interruption"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "latest_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SnapshotResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "non_endpoint_documents": {
            "title": "Non Endpoint Documents",
            "type": "integer"
          },
          "overall_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Overall Breakdown",
            "type": "object"
          },
          "read_in_full": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read In Full"
          },
          "rescan_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RescanReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "rows_unrecognized": {
            "title": "Rows Unrecognized",
            "type": "integer"
          },
          "scan_status": {
            "$ref": "#/components/schemas/ScanStatus"
          },
          "scanned_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scanned At"
          },
          "scanner_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scanner Version"
          },
          "section_rollup": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "title": "Section Rollup",
            "type": "object"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "top_issues": {
            "items": {
              "$ref": "#/components/schemas/IssueCount"
            },
            "title": "Top Issues",
            "type": "array"
          },
          "unread_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Unread Breakdown",
            "type": "object"
          },
          "unread_documents": {
            "title": "Unread Documents",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "label",
          "scan_status",
          "documents",
          "read_in_full",
          "docs_ok",
          "scanner_version",
          "scanned_at",
          "docs_changed",
          "rescan_reason",
          "overall_breakdown",
          "unread_breakdown",
          "unread_documents",
          "rows_unrecognized",
          "section_rollup",
          "error",
          "error_at",
          "active_snapshot",
          "latest_snapshot",
          "head_commit",
          "interruption",
          "top_issues",
          "non_endpoint_documents"
        ],
        "title": "ServiceDetailResponse",
        "type": "object"
      },
...
      "ServicesResponse": {
        "description": "The registry: matching rows plus the count behind every filter chip.",
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ServiceListItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "counts"
        ],
        "title": "ServicesResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Panel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scan/services": {
      "get": {
        "description": "The registry table: filtered rows plus the count behind every chip.\n\n``counts`` is computed with the search applied but the status filter\nignored, so the chips keep showing where the rest of the matches went.",
        "operationId": "list_services_api_scan_services_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "quality",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "rule",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Services"
      }
    },
    "/api/scan/services/{repo}": {
      "get": {
        "description": "Everything the service page shows about one service.",
        "operationId": "get_service_api_scan_services__repo__get",
        "parameters": [
          {
            "in": "path",
            "name": "repo",
            "required": true,
            "schema": {
              "title": "Repo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Service"
      }
    },
    "/api/scan/services/{repo}/documents": {
      "get": {
        "description": "One page of the active snapshot's API documents.\n\nPages that carry no scan status are not API documents; they are counted in\nthe service detail's ``non_endpoint_documents`` instead of being listed here\nunder a status they never had.\n\n``section`` and ``section_status`` together answer \"which documents have a\nfailed body\": both are required for the filter to apply, because either one\nalone would silently mean something else. ``issue`` answers the same\nquestion for a diagnostic code, and ``api_version`` narrows to one version\nof the API (``unversioned`` for the documents that name none).",
        "operationId": "list_documents_api_scan_services__repo__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "repo",
            "required": true,
            "schema": {
              "title": "Repo",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "section",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section"
            }
          },
          {
            "in": "query",
            "name": "section_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section Status"
            }
          },
          {
            "in": "query",
            "name": "issue",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issue"
            }
          },
          {
            "in": "query",
            "name": "api_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Documents"
      }
    },
    "/api/scan/services/{repo}/documents/{document_id}": {
      "get": {
        "description": "One document of the active snapshot, with its sections and parameters.",
        "operationId": "get_document_api_scan_services__repo__documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "repo",
            "required": true,
            "schema": {
              "title": "Repo",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document"
      }
    },
    "/api/scan/services/{repo}/export": {
      "get": {
        "description": "The active snapshot as a ``RepositoryScanResult`` - the scanner's own\ncontract, not a shape invented for this endpoint.\n\nRebuilt from the stored payloads and validated on the way out: if anything\nin the database no longer satisfies the contract, this fails loudly instead\nof handing out a file that only looks right.",
        "operationId": "export_snapshot_api_scan_services__repo__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "repo",
            "required": true,
            "schema": {
              "title": "Repo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Snapshot"
      }
    },
...
    "/api/scan/services/{repo}/snapshots": {
      "get": {
        "description": "A service's scan history, newest first.",
        "operationId": "list_snapshots_api_scan_services__repo__snapshots_get",
        "parameters": [
          {
            "in": "path",
            "name": "repo",
            "required": true,
            "schema": {
              "title": "Repo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Snapshots"
      }
    },
...
  }
}